refactor: move autonomous prompt workflow module
